+++ fearless_simd/src/lib.rs
@@ -77,6 +77,7 @@
 //! - `libm`: Use floating point implementations from [libm].
 //! - `safe_wrappers`: Include safe wrappers for (some) target feature specific intrinsics,
 //!   beyond the basic SIMD operations abstracted on all platforms.
+//! - `force_support_fallback`: Force scalar fallback, to be supported, even if your compilation target has a better baseline.
 //!
 //! At least one of `std` and `libm` is required; `std` overrides `libm`.
 //!
@@ -149,7 +150,25 @@ pub enum Level {
     /// Scalar fallback level, i.e. no supported SIMD features are to be used.
     ///
     /// This can be created with [`Level::fallback`].
-    // TODO: Allow not compiling this in (probably only on web, but maybe elsewhere?)
+    // We only want to compile the fallback implementation if:
+    // - We're on a supported architecture, but don't statically support the lowest alternative level; OR
+    // - We're on an unsupported architecture; OR
+    // - The fallback is forcibly enabled
+    #[cfg(any(
+        all(target_arch = "aarch64", not(target_feature = "neon")),
+        all(
+            any(target_arch = "x86", target_arch = "x86_64"),
+            not(target_feature = "sse4.2")
+        ),
+        all(target_arch = "wasm32", not(target_feature = "simd128")),
+        not(any(
+            target_arch = "x86",
+            target_arch = "x86_64",
+            target_arch = "aarch64",
+            target_arch = "wasm32"
+        )),
+        feature = "force_support_fallback"
+    ))]
     Fallback(Fallback),
     /// The Neon instruction set on 64 bit ARM.
     #[cfg(target_arch = "aarch64")]
@@ -158,7 +177,11 @@ pub enum Level {
     #[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
     WasmSimd128(WasmSimd128),
     /// The SSE4.2 instruction set on (32 and 64 bit) x86.
-    #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
+    // We don't need to support this if the compilation target definitely supports something better.
+    #[cfg(all(
+        any(target_arch = "x86", target_arch = "x86_64"),
+        not(all(target_feature = "avx2", target_feature = "fma"))
+    ))]
     Sse4_2(Sse4_2),
     /// The AVX2 and FMA instruction set on (32 and 64 bit) x86.
     #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
@@ -204,8 +227,6 @@ impl Level {
             // WASM always either has the SIMD feature compiled in or not.
             #[cfg(target_feature = "simd128")]
             return Level::WasmSimd128(WasmSimd128::new_unchecked());
-            #[cfg(not(target_feature = "simd128"))]
-            return Level::fallback();
         }
         #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
         {
@@ -214,11 +235,34 @@ impl Level {
             {
                 return unsafe { Level::Avx2(Avx2::new_unchecked()) };
             } else if std::arch::is_x86_feature_detected!("sse4.2") {
+                #[cfg(not(all(target_feature = "avx2", target_feature = "fma")))]
                 return unsafe { Level::Sse4_2(Sse4_2::new_unchecked()) };
             }
         }
-        #[cfg(not(target_arch = "wasm32"))]
-        Self::fallback()
+        #[cfg(any(
+            all(target_arch = "aarch64", not(target_feature = "neon")),
+            all(
+                any(target_arch = "x86", target_arch = "x86_64"),
+                not(target_feature = "sse4.2")
+            ),
+            all(target_arch = "wasm32", not(target_feature = "simd128")),
+            not(any(
+                target_arch = "x86",
+                target_arch = "x86_64",
+                target_arch = "aarch64",
+                target_arch = "wasm32"
+            )),
+        ))]
+        {
+            return Level::Fallback(Fallback::new());
+        }
+        #[allow(
+            unreachable_code,
+            reason = "`is_x86_feature_detected` or equivalents will have returned `true`, or Fallback was used."
+        )]
+        {
+            unreachable!()
+        }
     }
 
     /// Get the target feature level suitable for this run.
@@ -246,6 +290,10 @@ impl Level {
     #[cfg(target_arch = "aarch64")]
     #[inline]
     pub fn as_neon(self) -> Option<Neon> {
+        #[allow(
+            unreachable_patterns,
+            reason = "On machines which statically support `neon`, there is only one variant."
+        )]
         match self {
             Level::Neon(neon) => Some(neon),
             _ => None,
@@ -263,6 +311,10 @@ impl Level {
     #[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
     #[inline]
     pub fn as_wasm_simd128(self) -> Option<WasmSimd128> {
+        #[allow(
+            unreachable_patterns,
+            reason = "On machines which statically support `simd128`, there is only one variant."
+        )]
         match self {
             Level::WasmSimd128(simd128) => Some(simd128),
             _ => None,
@@ -281,7 +333,16 @@ impl Level {
     #[inline]
     pub fn as_sse4_2(self) -> Option<Sse4_2> {
         match self {
+            // Safety: The Avx2 struct represents the `avx2` and `fma` target features being enabled.
+            // The `avx2` target feature *also* implicitly enables the "sse4.2" target feature, which is
+            // the only target feature required to make our Sse4_2 token.
+            Level::Avx2(_avx) => unsafe { Some(Sse4_2::new_unchecked()) },
+            #[cfg(not(all(target_feature = "avx2", target_feature = "fma")))]
             Level::Sse4_2(sse42) => Some(sse42),
+            #[allow(
+                unreachable_patterns,
+                reason = "This arm is reachable on baseline x86/x86_64."
+            )]
             _ => None,
         }
     }
@@ -297,16 +358,86 @@ impl Level {
     #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
     #[inline]
     pub fn as_avx2(self) -> Option<Avx2> {
+        #[allow(
+            unreachable_patterns,
+            reason = "On machines which statically support `avx2`, there is only one variant."
+        )]
         match self {
             Level::Avx2(avx2) => Some(avx2),
             _ => None,
         }
     }
 
+    /// Get the strongest statically supported SIMD level.
+    ///
+    /// That is, if your compilation run ambiently declares that a target feature is enabled,
+    /// this method will take that into account.
+    /// In most cases, you should use [`Level::new`] or [`Level::try_detect`].
+    /// This method is mainly useful for libraries, where:
+    ///
+    /// 1) Your crate features request that you not use the standard library, i.e. doesn't enable
+    ///    your `"std"` crate feature reason (so you can't use [`Level::new`] and
+    ///    [`Level::try_detect`] returns `None`); AND
+    /// 2) Your caller does not provide a [`Level`]; AND
+    /// 3) The library doesn't want to panic when it can't find a SIMD level.
+    ///
+    /// Note that in these cases, the library should clearly inform the integrator
+    /// that it is using a fallback and so not getting optimal performance (e.g. by panicking if
+    /// `debug_assertions` are enabled, and emitting a log with the "error" level otherwise).
+    /// The messages given should also provide actionable fixes, such as pointing to the
+    /// entry-point which provides a `Level`, or your `"std"` feature.
+    ///
+    /// Note that this is unaffected by the `force-support-fallback` feature.
+    /// Instead, you should use [`Level::fallback`] if you require the fallback level.
+    pub const fn baseline() -> Self {
+        // TODO: How do we possibly test that this method works in all cases?
+        // Note that you can use the `check_targets.sh` script to at least ensure that it compiles in all reasonable cases.
+        #[cfg(not(any(
+            target_arch = "x86",
+            target_arch = "x86_64",
+            target_arch = "aarch64",
+            target_arch = "wasm32"
+        )))]
+        {
+            return Level::Fallback(Fallback::new());
+        }
+        #[cfg(target_arch = "aarch64")]
+        {
+            #[cfg(target_feature = "neon")]
+            return unsafe { Level::Neon(Neon::new_unchecked()) };
+            #[cfg(not(target_feature = "neon"))]
+            return Level::Fallback(Fallback::new());
+        }
+        #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
+        {
+            #[cfg(all(target_feature = "avx2", target_feature = "fma"))]
+            return unsafe { Level::Avx2(Avx2::new_unchecked()) };
+            #[cfg(all(
+                target_feature = "sse4.2",
+                not(all(target_feature = "avx2", target_feature = "fma"))
+            ))]
+            return unsafe { Level::Sse4_2(Sse4_2::new_unchecked()) };
+            #[cfg(not(target_feature = "sse4.2"))]
+            return Level::Fallback(Fallback::new());
+        }
+        #[cfg(target_arch = "wasm32")]
+        {
+            #[cfg(target_feature = "simd128")]
+            return Level::WasmSimd128(WasmSimd128::new_unchecked());
+            #[cfg(not(target_feature = "simd128"))]
+            return Level::Fallback(Fallback::new());
+        }
+    }
+
     /// Create a scalar fallback level, which uses no SIMD instructions.
     ///
-    /// This is primarily intended for tests; most users should prefer [`Level::new`].
+    /// This is primarily intended for tests; most users should prefer [`Level::new`] or [`Level::baseline`].
+    ///
+    /// Note that enabling the scalar fallback does *not* mean that the fallback branch will not
+    /// contain SIMD instructions. This is because the "ambient" compilation environment has SIMD
+    /// instructions available, which may be utilised by LLVM to auto-vectorise that path.
     #[inline]
+    #[cfg(feature = "force_support_fallback")]
     pub const fn fallback() -> Self {
         Self::Fallback(Fallback::new())
     }
@@ -328,50 +459,12 @@ impl Level {
     ///
     /// [enabled]: https://doc.rust-lang.org/reference/attributes/codegen.html#the-target_feature-attribute
     #[inline]
+    #[expect(
+        unreachable_patterns,
+        reason = "Level is `non_exhaustive`, but we are in the crate it's defined."
+    )]
     pub fn dispatch<W: WithSimd>(self, f: W) -> W::Output {
-        #[cfg(target_arch = "aarch64")]
-        #[target_feature(enable = "neon")]
-        #[inline]
-        fn dispatch_neon<W: WithSimd>(f: W, neon: Neon) -> W::Output {
-            f.with_simd(neon)
-        }
-
-        #[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
-        #[inline]
-        fn dispatch_simd128<W: WithSimd>(f: W, simd128: WasmSimd128) -> W::Output {
-            f.with_simd(simd128)
-        }
-
-        #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-        #[target_feature(enable = "sse4.2")]
-        #[inline]
-        fn dispatch_sse4_2<W: WithSimd>(f: W, sse4_2: Sse4_2) -> W::Output {
-            f.with_simd(sse4_2)
-        }
-
-        #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-        #[target_feature(enable = "avx2,fma")]
-        #[inline]
-        fn dispatch_avx2<W: WithSimd>(f: W, avx2: Avx2) -> W::Output {
-            f.with_simd(avx2)
-        }
-
-        #[inline]
-        fn dispatch_fallback<W: WithSimd>(f: W, fallback: Fallback) -> W::Output {
-            f.with_simd(fallback)
-        }
-
-        match self {
-            #[cfg(target_arch = "aarch64")]
-            Level::Neon(neon) => unsafe { dispatch_neon(f, neon) },
-            #[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
-            Level::WasmSimd128(simd128) => dispatch_simd128(f, simd128),
-            #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-            Level::Sse4_2(sse4_2) => unsafe { dispatch_sse4_2(f, sse4_2) },
-            #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-            Level::Avx2(avx2) => unsafe { dispatch_avx2(f, avx2) },
-            Level::Fallback(fallback) => dispatch_fallback(f, fallback),
-        }
+        dispatch!(self, simd => f.with_simd(simd))
     }
 }