+++ fearless_simd/src/macros.rs
@@ -3,117 +3,6 @@
 
 //! Macros publicly exported
 
-/// Defines a new function which dispatches to a SIMD-generic function, enabling the correct
-/// target features.
-///
-/// The `fn` token in the definition can be prefixed with a visibility (e.g. `pub`),
-/// to set the visibility of the outer function.
-/// We recommend that the implementation function remains private, and
-/// should only be called through the dispatch function.
-/// (The exact patterns for SIMD functions using Fearleess SIMD have not
-/// yet been designed/enumerated).
-///
-/// The implementation function (which is outside of this macro) *should* have the
-/// `#[inline(always)]` attribute.
-/// There are likely to be severe performance consequences if this is not the case, as
-/// Rust will be unable to inline SIMD intrinsics in that case.
-///
-/// The `fn` token in the definition can be prefixed with `unsafe`, to allow an unsafe inner function.
-/// The safety comment added by you in the call to  `simd_dispatch` the function must have
-/// the preconditions required to call the inner function.
-///
-/// # Examples
-///
-/// ```rust
-/// use fearless_simd::{Simd, simd_dispatch};
-///
-/// #[inline(always)]
-/// fn sigmoid_impl<S: Simd>(simd: S, x: &[f32], out: &mut [f32]) { /* ... */ }
-///
-/// simd_dispatch!(fn sigmoid(level, x: &[f32], out: &mut [f32]) = sigmoid_impl);
-/// ```
-///
-/// The signature of the generated function will be:
-///
-/// ```rust
-/// use fearless_simd::Level;
-/// fn sigmoid(level: Level, x: &[f32], out: &mut [f32]) { /* ... */ }
-/// ```
-#[macro_export]
-#[deprecated = "use dispatch!(level, simd => operation) instead"]
-macro_rules! simd_dispatch {
-    (
-        $( #[$meta:meta] )* $vis:vis
-        unsafe fn $func:ident ( level $( , $arg:ident : $ty:ty $(,)? )* ) $( -> $ret:ty )?
-        = $inner:ident
-    ) => {
-        simd_dispatch!{@impl => $(#[$meta])* $vis (unsafe) fn $func (level, $(,$arg:$ty,)*) $(->$ret)? = $inner}
-    };
-    (
-        $( #[$meta:meta] )* $vis:vis
-        fn $func:ident ( level $( , $arg:ident : $ty:ty $(,)? )* ) $( -> $ret:ty )?
-        = $inner:ident
-    ) => {
-        simd_dispatch!{@impl => $(#[$meta])* $vis () fn $func (level $(,$arg:$ty)*) $(->$ret)? = $inner}
-    };
-    (
-        @impl => $( #[$meta:meta] )* $vis:vis
-        ($($unsafe: ident)?) fn $func:ident ( level $( , $arg:ident : $ty:ty $(,)? )* ) $( -> $ret:ty )?
-        = $inner:ident
-    ) => {
-        $( #[$meta] )* $vis
-        $($unsafe)? fn $func(level: $crate::Level $(, $arg: $ty )*) $( -> $ret )? {
-            #[cfg(target_arch = "aarch64")]
-            #[target_feature(enable = "neon")]
-            #[inline]
-            $($unsafe)? fn inner_neon(neon: $crate::aarch64::Neon $( , $arg: $ty )* ) $( -> $ret )? {
-                $($unsafe)? {
-                    $inner( neon $( , $arg )* )
-                }
-            }
-            #[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
-            #[inline]
-            $($unsafe)? fn inner_wasm_simd128(simd128: $crate::wasm32::WasmSimd128 $( , $arg: $ty )* ) $( -> $ret )? {
-                $($unsafe)? {
-                    $inner( simd128 $( , $arg )* )
-                }
-            }
-            #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-            #[target_feature(enable = "sse4.2")]
-            #[inline]
-            $($unsafe)? fn inner_sse4_2(sse4_2: $crate::x86::Sse4_2 $( , $arg: $ty )* ) $( -> $ret )? {
-                $($unsafe)? {
-                    $inner( sse4_2 $( , $arg )* )
-                }
-            }
-            #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-            #[target_feature(enable = "avx2,fma")]
-            #[inline]
-            $($unsafe)? fn inner_avx2(avx2: $crate::x86::Avx2 $( , $arg: $ty )* ) $( -> $ret )? {
-                $($unsafe)? {
-                    $inner( avx2 $( , $arg )* )
-                }
-            }
-            match level {
-                $crate::Level::Fallback(fb) => {
-                    $($unsafe)? {
-                        $inner(fb $( , $arg )* )
-                    }
-                },
-                #[cfg(target_arch = "aarch64")]
-                $crate::Level::Neon(neon) => unsafe { inner_neon (neon $( , $arg )* ) }
-                #[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
-                $crate::Level::WasmSimd128(wasm) => unsafe { inner_wasm_simd128 (wasm $( , $arg )* ) }
-                #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-                $crate::Level::Sse4_2(sse4_2) => unsafe { inner_sse4_2(sse4_2 $( , $arg)* ) }
-                #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-                $crate::Level::Avx2(avx2) => unsafe { inner_avx2(avx2 $( , $arg)* ) }
-                _ => unreachable!()
-            }
-        }
-    };
-}
-
 /// Access the applicable [`Simd`] for a given `level`, and perform an operation using it.
 ///
 /// This macro is the root of how any explicitly written SIMD functions in this crate are
@@ -158,24 +47,19 @@ macro_rules! simd_dispatch {
 /// [`Simd`]: crate::Simd
 #[macro_export]
 macro_rules! dispatch {
-    ($level:expr, $simd:pat => $op:expr) => {{
+    // This falls through to the next branch, but with `forced_fallback_arm` turned into a boolean literal
+    // indicating whether or not the `force_support_fallback` crate feature is enabled.
+    ($level:expr, $simd:pat => $op:expr) => {{ $crate::internal_unstable_dispatch_inner!($level, $simd => $op) }};
+    (@impl $level:expr, $simd:pat => $op:expr; $forced_fallback_arm: literal) => {{
         /// Convert the `Simd` value into an `impl Simd`, which enforces that
         /// it is correctly handled.
+        // TODO: Just make into a `pub` function in fearless_simd itself?
         #[inline(always)]
         fn launder<S: $crate::Simd>(x: S) -> impl $crate::Simd {
             x
         }
 
         match $level {
-            $crate::Level::Fallback(fb) => {
-                let $simd = launder(fb);
-                // This vectorize call does nothing, but it is reasonable to be consistent here.
-                $crate::Simd::vectorize(
-                    fb,
-                    #[inline(always)]
-                    || $op,
-                )
-            }
             #[cfg(target_arch = "aarch64")]
             $crate::Level::Neon(neon) => {
                 let $simd = launder(neon);
@@ -194,7 +78,11 @@ macro_rules! dispatch {
                     || $op,
                 )
             }
-            #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
+            // This fallthrough logic is documented at the definition site of `Level`.
+            #[cfg(all(
+                any(target_arch = "x86", target_arch = "x86_64"),
+                not(all(target_feature = "avx2", target_feature = "fma"))
+            ))]
             $crate::Level::Sse4_2(sse4_2) => {
                 let $simd = launder(sse4_2);
                 $crate::Simd::vectorize(
@@ -212,9 +100,60 @@ macro_rules! dispatch {
                     || $op,
                 )
             }
+            #[cfg(any(
+                all(target_arch = "aarch64", not(target_feature = "neon")),
+                all(
+                    any(target_arch = "x86", target_arch = "x86_64"),
+                    not(target_feature = "sse4.2")
+                ),
+                all(target_arch = "wasm32", not(target_feature = "simd128")),
+                not(any(
+                    target_arch = "x86",
+                    target_arch = "x86_64",
+                    target_arch = "aarch64",
+                    target_arch = "wasm32"
+                )),
+                $forced_fallback_arm
+            ))]
+            $crate::Level::Fallback(fb) => {
+                let $simd = launder(fb);
+                // This vectorize call does nothing, but it is reasonable to be consistent here.
+                $crate::Simd::vectorize(
+                    fb,
+                    #[inline(always)]
+                    || $op,
+                )
+            }
             _ => unreachable!(),
         }
     }};
+}
+
+// This macro turns whether the `force_support_fallback` macro is enabled into a boolean literal
+// in `dispatch`, which allows it to be used correctly cross-crate.
+// This trickery is required because macros are expanded in the context of the calling crate, including for
+// evaluating `cfg`s.
+
+/// Implementation detail of [`crate::dispatch`]; this is not public API.
+#[macro_export]
+#[doc(hidden)]
+#[cfg(feature = "force_support_fallback")]
+macro_rules! internal_unstable_dispatch_inner {
+    ($level:expr, $simd:pat => $op:expr) => {
+        $crate::dispatch!(
+            @impl $level, $simd => $op; true
+        )
+    };
+}
+
+/// Implementation detail of [`crate::dispatch`]; this is not public API.
+#[macro_export]
+#[doc(hidden)]
+#[cfg(not(feature = "force_support_fallback"))]
+macro_rules! internal_unstable_dispatch_inner {
+    ($level:expr, $simd:pat => $op:expr) => {
+        $crate::dispatch!(@impl $level, $simd => $op; false)
+    };
 }
 
 #[cfg(test)]